Check container is running after docker start.
This attempts to fix the problem with Buildbot tests giving:
docker.errors.APIError: 500 Server Error: Internal Server Error
	("You cannot remove a running container. Stop the container
	before attempting removal or use -f")

Theory is that 'docker start' returns before the container is fully
in Running state.  A subsequent 'docker stop' sees that Running is
False and indicates that the container is stopped without actually
stopping it.  A subsequent 'docker rm' then gives the above error.